requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
nltk
wordcloud
scikit-learn
google-generativeai
fastapi
uvicorn